See Description
Lots of changes to the ImGui Interface.

Created a base namespace for the managing of multiple ImGui Windows for ease of creating and rendering.

Created MOTD interface for Imgui that pulls the message of the day from the website using Libcurl

Recfactored Tag fixes that were spread across various different namespaces into a singular namespace TagFixes.

Refactored code that enables death cam target switching for MK

Added new options to Custom Resolutions, as well as adding one that adds your monitors resolution.
