+++ xlive/H2MOD.cpp
@@ -27,7 +27,7 @@
 #include "H2MOD/Tags/MetaExtender.h"
 #include "H2MOD/Modules/MainLoopPatches/UncappedFPS2/UncappedFPS2.h"
 #include "H2MOD/Modules/Input/ControllerInput.h"
-#include "H2MOD/Modules/SettingHooks/ButtonSettings.h"
+#include "H2MOD/Modules/TagFixes/TagFixes.h"
 
 H2MOD* h2mod = new H2MOD();
 GunGame* gunGame = new GunGame();
@@ -832,7 +832,6 @@ bool __cdecl OnMapLoad(Blam::EngineDefinitions::game_engine_settings* engine_set
 
 		H2Tweaks::toggleAiMp(false);
 		H2Tweaks::toggleUncappedCampaignCinematics(false);
-		SettingsMenus::Controller::ButtonLayout::Initialize();
 		MetaExtender::free_tag_blocks();
 		return result;
 	}
@@ -1640,6 +1639,7 @@ void H2MOD::Initialize()
 		MouseInput::Initialize();
 		KeyboardInput::Initialize();
 		ControllerInput::Initialize();
+		TagFixes::Initalize();
 		if (H2Config_discord_enable && H2GetInstanceId() == 1) {
 			// Discord init
 			DiscordInterface::SetDetails("Startup");
