+++ xlive/H2MOD/GUI/imgui_integration/imgui_handler.h
@@ -3,6 +3,22 @@
 
 namespace imgui_handler
 {
+	struct s_imgui_window
+	{
+		std::string name;
+		bool DoRender;
+		std::function<void(bool*)> RenderFunc;
+		std::function<void()> OpenFunc;
+		std::function<void()> CloseFunc;
+		s_imgui_window(std::string name, bool doRender, std::function<void(bool*)> renderFunc, std::function<void()> openFunc, std::function<void()> closeFunc)
+		{
+			this->name = name;
+			this->DoRender = doRender;
+			this->RenderFunc = renderFunc;
+			this->OpenFunc = openFunc;
+			this->CloseFunc = closeFunc;
+		}
+	};
 	HWND get_HWND();
 	bool CanDrawImgui();
 	void DrawImgui();
@@ -15,6 +31,16 @@ namespace imgui_handler
 	float WidthPercentage(float percent);
 	void TextVerticalPad(char* label, float amount);
 
-	void AdvancedSettings(bool* p_open);
-	void MessageOfTheDay(bool* p_open);
+
+	namespace MOTD {
+		void Render(bool* p_open);
+		void Open();
+		void Close();
+	}
+	namespace AdvancedSettings
+	{
+		void Render(bool* p_open);
+		void Open();
+		void Close();
+	}
 }
