+++ xlive/H2MOD/Modules/Input/ControllerInput.h
@@ -1,6 +1,7 @@
 #pragma once
 namespace ControllerInput
 {
+	char* __cdecl get_controller_input(__int16 index);
 	void SetSensitiviy(float value);
 	void Initialize();
 }
