+++ xlive/H2MOD/Modules/TagFixes/TagFixes.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace TagFixes
+{
+	void OnMapLoad();
+	void Initalize();
+}
