+++ xlive/H2MOD/Modules/Tweaks/Tweaks.cpp
@@ -713,57 +713,6 @@ class test_engine : public c_game_engine_base
 };
 test_engine g_test_engine;
 
-
-void fix_shader_template_nvidia(const std::string &template_name, const std::string &bitmap_name, size_t bitmap_idx)
-{
-	datum bitmap_to_fix    = tags::find_tag(blam_tag::tag_group_type::bitmap, bitmap_name);
-	datum borked_template  = tags::find_tag(blam_tag::tag_group_type::shadertemplate, template_name);
-
-	LOG_DEBUG_FUNC("bitmap {0}, borked_template {1}", bitmap_to_fix.data, borked_template.data);
-
-	if (bitmap_to_fix.IsNull() || borked_template.IsNull())
-		return;
-
-	LOG_DEBUG_FUNC("Fixing: template {}, bitmap {}", template_name, bitmap_name);
-
-	tags::ilterator shaders(blam_tag::tag_group_type::shader);
-	while (!shaders.next().IsNull())
-	{
-		auto *shader = LOG_CHECK(tags::get_tag<blam_tag::tag_group_type::shader, shader_definition>(shaders.m_datum));
-		if (shader && shader->shader_template.TagIndex == borked_template && LOG_CHECK(shader->postprocessDefinition.size > 0))
-		{
-			LOG_DEBUG_FUNC("shader {} has borked template", tags::get_tag_name(shaders.m_datum));
-			auto *post_processing = shader->postprocessDefinition[0];
-			if (LOG_CHECK(post_processing->bitmaps.size >= (bitmap_idx + 1)))
-			{
-				auto *bitmap_block = post_processing->bitmaps[bitmap_idx];
-				if (bitmap_block->bitmapGroup == bitmap_to_fix)
-				{
-					LOG_DEBUG_FUNC("Nulled bitmap {}", bitmap_idx);
-					bitmap_block->bitmapGroup = datum::Null;
-				}
-			}
-		}
-	}
-}
-
-void fix_shaders_nvidia()
-{
-	if (h2mod->Server) return;
-
-	fix_shader_template_nvidia(
-		"shaders\\shader_templates\\opaque\\tex_bump_alpha_test_single_pass", 
-		"shaders\\default_bitmaps\\bitmaps\\alpha_white",
-		4
-	);
-
-	fix_shader_template_nvidia(
-		"shaders\\shader_templates\\opaque\\tex_bump_alpha_test",
-		"shaders\\default_bitmaps\\bitmaps\\gray_50_percent",
-		1
-	);
-}
-
 void InitH2Tweaks() {
 	postConfig();
 
@@ -831,7 +780,7 @@ void InitH2Tweaks() {
 		WriteJmpTo(h2mod->GetAddress(0x7E43), WinMain);
 		WriteJmpTo(h2mod->GetAddress(0x39EA2), is_remote_desktop);
 
-		tags::on_map_load(fix_shaders_nvidia);
+		//tags::on_map_load(fix_shaders_nvidia);
 		tags::on_map_load(c_xbox_live_task_progress_menu::ApplyPatches);
 
 		// disable cloth debugging that writes to cloth.txt
