+++ xlive/Project_Cartographer.vcxproj
@@ -293,8 +293,8 @@
     <ClCompile Include="H2MOD\Modules\Networking\NetworkObserver\NetworkObserver.cpp" />
     <ClCompile Include="H2MOD\Modules\Networking\NetworkSession\NetworkSession.cpp" />
     <ClCompile Include="H2MOD\Modules\Networking\NetworkStats\NetworkStats.cpp" />
-    <ClCompile Include="H2MOD\Modules\SettingHooks\ButtonSettings.cpp" />
     <ClCompile Include="H2MOD\Modules\Stats\StatsHandler.cpp" />
+    <ClCompile Include="H2MOD\Modules\TagFixes\TagFixes.cpp" />
     <ClCompile Include="H2MOD\Modules\UI\Pregame_Lobby.cpp" />
     <ClCompile Include="H2MOD\Modules\UI\XboxLiveTaskProgress.cpp" />
     <ClCompile Include="H2MOD\Tags\MetaExtender.cpp" />
@@ -529,8 +529,8 @@
     <ClInclude Include="H2MOD\Modules\Networking\NetworkSession\NetworkSession.h" />
     <ClInclude Include="H2MOD\Modules\Networking\NetworkStats\NetworkStats.h" />
     <ClInclude Include="H2MOD\Modules\ServerConsole\ServerConsole.h" />
-    <ClInclude Include="H2MOD\Modules\SettingHooks\ButtonSettings.h" />
     <ClInclude Include="H2MOD\Modules\Stats\StatsHandler.h" />
+    <ClInclude Include="H2MOD\Modules\TagFixes\TagFixes.h" />
     <ClInclude Include="H2MOD\Modules\UI\Pregame_Lobby.h" />
     <ClInclude Include="H2MOD\Modules\UI\XboxLiveTaskProgress.h" />
     <ClInclude Include="H2MOD\Tags\MetaExtender.h" />
