+++ xlive/Project_Cartographer.vcxproj.filters
@@ -103,7 +103,7 @@
     <ClCompile Include="H2MOD\GUI\imgui_integration\imgui_handler.cpp" />
     <ClCompile Include="H2MOD\GUI\imgui_integration\AdvancedSettings.cpp" />
     <ClCompile Include="H2MOD\GUI\imgui_integration\MOTD.cpp" />
-    <ClCompile Include="H2MOD\Modules\SettingHooks\ButtonSettings.cpp" />
+    <ClCompile Include="H2MOD\Modules\TagFixes\TagFixes.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="3rdparty\discord\discord_register.h" />
@@ -351,7 +351,7 @@
     <ClInclude Include="H2MOD\Modules\Input\ControllerInput.h" />
     <ClInclude Include="H2MOD\GUI\imgui_integration\imgui_handler.h" />
     <ClInclude Include="Blam\Cache\TagGroups\s_user_interface_screen_widget_definition.hpp" />
-    <ClInclude Include="H2MOD\Modules\SettingHooks\ButtonSettings.h" />
+    <ClInclude Include="H2MOD\Modules\TagFixes\TagFixes.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="Resource.rc" />
